+++ widget/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @betagouv/les-communs-widget
 
+## 0.3.1
+
+### Minor Changes
+
+- removing logs from widget
+
+## 0.3.0
+
+### Patch Changes
+
+- make widget compatible without providing a project id from les Communs
+
 ## 0.2.6
 
 ### Minor Changes